Moved extensions to dedicated folder.
