+++ Package.swift
@@ -1,10 +1,11 @@
-// swift-tools-version:5.9
+// swift-tools-version:5.7
 import PackageDescription
 
 let package = Package(
   name: "PovioKit",
   platforms: [
-    .iOS(.v13), .macOS(.v12)
+    .iOS(.v13),
+    .macOS(.v12)
   ],
   products: [
     .library(
